requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg-resources==0.0.0
pymongo==3.11.4
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
Werkzeug==2.0.1
